Move dynamic code generator into mod.rs
