+++ crates/cli/src/codegen/mod.rs
@@ -36,9 +36,6 @@
 mod builder;
 pub(crate) use builder::*;
 
-mod dynamic;
-pub(crate) use dynamic::*;
-
 mod r#static;
 pub(crate) use r#static::*;
 
@@ -46,8 +43,11 @@ mod transform;
 
 mod exports;
 pub(crate) use exports::*;
+use transform::SourceCodeSection;
+use walrus::{DataId, DataKind, FunctionBuilder, FunctionId, LocalId, MemoryId, Module, ValType};
 
-use crate::JS;
+use crate::{providers::Provider, JS};
+use anyhow::Result;
 
 pub(crate) enum CodeGenType {
     /// Static code generation.
@@ -61,4 +61,325 @@ pub(crate) enum CodeGenType {
 pub(crate) trait CodeGen {
     /// Generate Wasm from a given JS source.
     fn generate(&mut self, source: &JS) -> anyhow::Result<Vec<u8>>;
+}
+
+/// Identifiers used by the generated dynamically linkable module.
+// This is an internal detail of this module.
+pub(crate) struct Identifiers {
+    canonical_abi_realloc: FunctionId,
+    eval_bytecode: FunctionId,
+    invoke: FunctionId,
+    memory: MemoryId,
+}
+
+impl Identifiers {
+    fn new(
+        canonical_abi_realloc: FunctionId,
+        eval_bytecode: FunctionId,
+        invoke: FunctionId,
+        memory: MemoryId,
+    ) -> Self {
+        Self {
+            canonical_abi_realloc,
+            eval_bytecode,
+            invoke,
+            memory,
+        }
+    }
+}
+
+/// Helper struct to keep track of bytecode metadata.
+// This is an internal detail of this module.
+pub(crate) struct BytecodeMetadata {
+    ptr: LocalId,
+    len: i32,
+    data_section: DataId,
+}
+
+impl BytecodeMetadata {
+    fn new(ptr: LocalId, len: i32, data_section: DataId) -> Self {
+        Self {
+            ptr,
+            len,
+            data_section,
+        }
+    }
+}
+
+/// Dynamic Code Generation.
+pub(crate) struct DynamicGenerator {
+    /// Provider to use.
+    pub provider: Provider,
+    /// JavaScript function exports.
+    function_exports: Exports,
+    /// Whether to embed the compressed JS source in the generated module.
+    pub source_compression: bool,
+    /// WIT options for code generation.
+    pub wit_opts: WitOptions,
+}
+
+impl DynamicGenerator {
+    /// Creates a new [`DynamicGenerator`].
+    pub fn new() -> Self {
+        Self {
+            source_compression: true,
+            provider: Provider::Default,
+            function_exports: Default::default(),
+            wit_opts: Default::default(),
+        }
+    }
+
+    /// Generate the starting module.
+    fn generate_initial_module(&self) -> Module {
+        Module::with_config(transform::module_config())
+    }
+
+    /// Resolve identifiers for functions and memory.
+    pub fn resolve_identifiers(&self, module: &mut Module) -> Result<Identifiers> {
+        let import_namespace = self.provider.import_namespace()?;
+        let canonical_abi_realloc_type = module.types.add(
+            &[ValType::I32, ValType::I32, ValType::I32, ValType::I32],
+            &[ValType::I32],
+        );
+        let (canonical_abi_realloc_fn_id, _) = module.add_import_func(
+            &import_namespace,
+            "canonical_abi_realloc",
+            canonical_abi_realloc_type,
+        );
+
+        let eval_bytecode_type = module.types.add(&[ValType::I32, ValType::I32], &[]);
+        let (eval_bytecode_fn_id, _) =
+            module.add_import_func(&import_namespace, "eval_bytecode", eval_bytecode_type);
+
+        let invoke_type = module.types.add(
+            &[ValType::I32, ValType::I32, ValType::I32, ValType::I32],
+            &[],
+        );
+        let (invoke_fn_id, _) =
+            module.add_import_func(&self.provider.import_namespace()?, "invoke", invoke_type);
+
+        let (memory_id, _) =
+            module.add_import_memory(&import_namespace, "memory", false, false, 0, None, None);
+
+        Ok(Identifiers::new(
+            canonical_abi_realloc_fn_id,
+            eval_bytecode_fn_id,
+            invoke_fn_id,
+            memory_id,
+        ))
+    }
+
+    /// Generate the main function.
+    fn generate_main(
+        &self,
+        module: &mut Module,
+        js: &JS,
+        imports: &Identifiers,
+    ) -> Result<BytecodeMetadata> {
+        let bytecode = js.compile(&self.provider)?;
+        let bytecode_len: i32 = bytecode.len().try_into()?;
+        let bytecode_data = module.data.add(DataKind::Passive, bytecode);
+
+        let mut main = FunctionBuilder::new(&mut module.types, &[], &[]);
+        let bytecode_ptr_local = module.locals.add(ValType::I32);
+        main.func_body()
+            // Allocate memory in javy_quickjs_provider for bytecode array.
+            .i32_const(0) // orig ptr
+            .i32_const(0) // orig size
+            .i32_const(1) // alignment
+            .i32_const(bytecode_len) // new size
+            .call(imports.canonical_abi_realloc)
+            // Copy bytecode array into allocated memory.
+            .local_tee(bytecode_ptr_local) // save returned address to local and set as dest addr for mem.init
+            .i32_const(0) // offset into data segment for mem.init
+            .i32_const(bytecode_len) // size to copy from data segment
+            // top-2: dest addr, top-1: offset into source, top-0: size of memory region in bytes.
+            .memory_init(imports.memory, bytecode_data)
+            // Evaluate bytecode.
+            .local_get(bytecode_ptr_local) // ptr to bytecode
+            .i32_const(bytecode_len)
+            .call(imports.eval_bytecode);
+        let main = main.finish(vec![], &mut module.funcs);
+
+        module.exports.add("_start", main);
+        Ok(BytecodeMetadata::new(
+            bytecode_ptr_local,
+            bytecode_len,
+            bytecode_data,
+        ))
+    }
+
+    /// Generate function exports.
+    fn generate_exports(
+        &self,
+        module: &mut Module,
+        identifiers: &Identifiers,
+        bc_metadata: &BytecodeMetadata,
+    ) -> Result<()> {
+        if !self.function_exports.is_empty() {
+            let fn_name_ptr_local = module.locals.add(ValType::I32);
+            for export in &self.function_exports {
+                // For each JS function export, add an export that copies the name of the function into memory and invokes it.
+                let js_export_bytes = export.js.as_bytes();
+                let js_export_len: i32 = js_export_bytes.len().try_into().unwrap();
+                let fn_name_data = module.data.add(DataKind::Passive, js_export_bytes.to_vec());
+
+                let mut export_fn = FunctionBuilder::new(&mut module.types, &[], &[]);
+                export_fn
+                    .func_body()
+                    // Copy bytecode.
+                    .i32_const(0) // orig ptr
+                    .i32_const(0) // orig len
+                    .i32_const(1) // alignment
+                    .i32_const(bc_metadata.len) // size to copy
+                    .call(identifiers.canonical_abi_realloc)
+                    .local_tee(bc_metadata.ptr)
+                    .i32_const(0) // offset into data segment
+                    .i32_const(bc_metadata.len) // size to copy
+                    .memory_init(identifiers.memory, bc_metadata.data_section) // copy bytecode into allocated memory
+                    .data_drop(bc_metadata.data_section)
+                    // Copy function name.
+                    .i32_const(0) // orig ptr
+                    .i32_const(0) // orig len
+                    .i32_const(1) // alignment
+                    .i32_const(js_export_len) // new size
+                    .call(identifiers.canonical_abi_realloc)
+                    .local_tee(fn_name_ptr_local)
+                    .i32_const(0) // offset into data segment
+                    .i32_const(js_export_len) // size to copy
+                    .memory_init(identifiers.memory, fn_name_data) // copy fn name into allocated memory
+                    .data_drop(fn_name_data)
+                    // Call invoke.
+                    .local_get(bc_metadata.ptr)
+                    .i32_const(bc_metadata.len)
+                    .local_get(fn_name_ptr_local)
+                    .i32_const(js_export_len)
+                    .call(identifiers.invoke);
+                let export_fn = export_fn.finish(vec![], &mut module.funcs);
+                module.exports.add(&export.wit, export_fn);
+            }
+        }
+        Ok(())
+    }
+}
+
+impl CodeGen for DynamicGenerator {
+    // Run the calling code with the `dump_wat` feature enabled to print the WAT to stdout
+    //
+    // For the example generated WAT, the `bytecode_len` is 137
+    // (module
+    //    (type (;0;) (func))
+    //    (type (;1;) (func (param i32 i32)))
+    //    (type (;2;) (func (param i32 i32 i32 i32)))
+    //    (type (;3;) (func (param i32 i32 i32 i32) (result i32)))
+    //    (import "javy_quickjs_provider_v2" "canonical_abi_realloc" (func (;0;) (type 3)))
+    //    (import "javy_quickjs_provider_v2" "eval_bytecode" (func (;1;) (type 1)))
+    //    (import "javy_quickjs_provider_v2" "memory" (memory (;0;) 0))
+    //    (import "javy_quickjs_provider_v2" "invoke" (func (;2;) (type 2)))
+    //    (func (;3;) (type 0)
+    //      (local i32 i32)
+    //      i32.const 0
+    //      i32.const 0
+    //      i32.const 1
+    //      i32.const 137
+    //      call 0
+    //      local.tee 0
+    //      i32.const 0
+    //      i32.const 137
+    //      memory.init 0
+    //      data.drop 0
+    //      i32.const 0
+    //      i32.const 0
+    //      i32.const 1
+    //      i32.const 3
+    //      call 0
+    //      local.tee 1
+    //      i32.const 0
+    //      i32.const 3
+    //      memory.init 1
+    //      data.drop 1
+    //      local.get 0
+    //      i32.const 137
+    //      local.get 1
+    //      i32.const 3
+    //      call 2
+    //    )
+    //    (func (;4;) (type 0)
+    //      (local i32)
+    //      i32.const 0
+    //      i32.const 0
+    //      i32.const 1
+    //      i32.const 137
+    //      call 0
+    //      local.tee 0
+    //      i32.const 0
+    //      i32.const 137
+    //      memory.init 0
+    //      local.get 0
+    //      i32.const 137
+    //      call 1
+    //    )
+    //    (export "_start" (func 4))
+    //    (export "foo" (func 3))
+    //    (data (;0;) "\02\05\18function.mjs\06foo\0econsole\06log\06bar\0f\bc\03\00\01\00\00\be\03\00\00\0e\00\06\01\a0\01\00\00\00\03\01\01\1a\00\be\03\00\01\08\ea\05\c0\00\e1)8\e0\00\00\00B\e1\00\00\00\04\e2\00\00\00$\01\00)\bc\03\01\04\01\00\07\0a\0eC\06\01\be\03\00\00\00\03\00\00\13\008\e0\00\00\00B\e1\00\00\00\04\df\00\00\00$\01\00)\bc\03\01\02\03]")
+    //    (data (;1;) "foo")
+    //  )
+    fn generate(&mut self, js: &JS) -> Result<Vec<u8>> {
+        if self.wit_opts.defined() {
+            self.function_exports = exports::process_exports(
+                js,
+                self.wit_opts.unwrap_path(),
+                self.wit_opts.unwrap_world(),
+            )?;
+        }
+
+        let mut module = self.generate_initial_module();
+        let identifiers = self.resolve_identifiers(&mut module)?;
+        let bc_metadata = self.generate_main(&mut module, js, &identifiers)?;
+        self.generate_exports(&mut module, &identifiers, &bc_metadata)?;
+
+        transform::add_producers_section(&mut module.producers);
+        if !self.source_compression {
+            module.customs.add(SourceCodeSection::uncompressed(js)?);
+        } else {
+            module.customs.add(SourceCodeSection::compressed(js)?);
+        }
+
+        let wasm = module.emit_wasm();
+        print_wat(&wasm)?;
+        Ok(wasm)
+    }
+}
+
+#[cfg(feature = "dump_wat")]
+fn print_wat(wasm_binary: &[u8]) -> Result<()> {
+    println!(
+        "Generated WAT: \n{}",
+        wasmprinter::print_bytes(&wasm_binary)?
+    );
+    Ok(())
+}
+
+#[cfg(not(feature = "dump_wat"))]
+fn print_wat(_wasm_binary: &[u8]) -> Result<()> {
+    Ok(())
+}
+
+#[cfg(test)]
+mod test {
+    use crate::providers::Provider;
+
+    use super::DynamicGenerator;
+    use super::WitOptions;
+    use anyhow::Result;
+
+    #[test]
+    fn default_values() -> Result<()> {
+        let gen = DynamicGenerator::new();
+        assert!(gen.source_compression);
+        assert!(matches!(gen.provider, Provider::Default));
+        assert_eq!(gen.wit_opts, WitOptions::default());
+
+        Ok(())
+    }
 }
